# src/main/resources/swagger.yaml
openapi: 3.0.0
info:
  title: CoActivity API
  description: API test
  version: 1.0.0

#servers:
#  - url: 'http://coactivity_app:8080/api/v1'
...
paths:
...
  /auth/register:
    post:
      tags:
        - Auth
      summary: Регистрация (после смены профиля)
      description: Создание нового аккаунта
      operationId: registerNewUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - email
                - password
              properties:
                name:
                  type: string
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '201':
          description: Регистрация успешна, новый пользователь создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Ошибка валидации или email занят
  /auth/login:
    post:
      tags:
        - Auth
      summary: Вход в аккаунт (после смены профиля)
      description: Вход в уже существующий аккаунт
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: JWT токен доступа
                  userId:
                    type: integer
                    description: ID пользователя
        '401':
          description: Неверный логин или пароль

  # --- USERS ---
...
components:
  schemas:
...
    User:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        id:
          type: integer
        passwordHash:
          type: string
        rooms:
          type: array
          items:
            type: integer
        interests:
          type: array
          items:
            type: integer
        feedbacks:
          type: array
          items:
            type: integer
        feedbacksAuthor:
          type: array
          items:
            type: integer
        createdAt:
          type: string
          format: date-time
...